User authentication

+++ studybud/base/views.py
@@ -1,6 +1,13 @@
 from django.shortcuts import render, redirect
-from .models import Room
+from django.http import HttpResponse
+from django.contrib import messages
+from django.db.models import Q
+from django.contrib.auth.decorators import login_required
+from .models import Room, Topic
 from .forms import RoomForm
+from django.contrib.auth.forms import UserCreationForm
+from django.contrib.auth.models import User
+from django.contrib.auth import authenticate, login, logout
 
 # Create your views here.
 
@@ -14,11 +21,51 @@ from .forms import RoomForm
 #     {'id': 7, 'name': 'Lets learn Django'},
 # ]
 
+def login_page(request):
+    page='login'
+    if request.user.is_authenticated:
+        return redirect('home')
+    
+    # Check if the method is POST
+    if request.method == 'POST':
+        username = request.POST.get('username')
+        password = request.POST.get('password')
+        
+        # Check if the user exists in the database
+        try:
+            user = User.objects.get(username=username)
+        except:
+            messages.error(request, 'The user does not exist')
+        
+        # Authenticate User if there's user
+        user = authenticate(request, username=username, password=password)
+        
+        # Check if the user is authenticated, log in
+        if user is not None:
+            login(request, user)
+            return redirect('home')
+        else:
+            messages.error(request, 'The username of password does not exist')
+    context = {'page': page}
+    return render(request, 'base/login_register.html', context)
+
+def logout_user(request):
+    logout(request)
+    return redirect('home')
+
+def register_user(request):
+    page = 'register'
+    form = UserCreationForm()
+    context = {'page': page, 'form': form}
+    return render(request, 'base/login_register.html', context)
 
 def home(request):
     # Query the database to get all rooms
-    rooms = Room.objects.all()
-    context = {'rooms': rooms}
+    q = request.GET.get('q') if request.GET.get('q') != None else ''
+    rooms = Room.objects.filter(Q(topic__name__icontains = q) | Q(name__icontains = q) | Q(description__icontains = q)) 
+    room_count = rooms.count()
+    topics = Topic.objects.all()
+    context = {'rooms': rooms, 'topics': topics, 'room_count': room_count}
     return render(request, 'base/home.html', context)
 
 
@@ -34,6 +81,7 @@ def room(request, pk):
 
 
 # Create Room
+@login_required(login_url='login')
 def create_room(request):
     # Create an instance of form class (make a form object)
     form = RoomForm()
@@ -47,14 +95,23 @@ def create_room(request):
     return render(request, 'base/room_form.html', context)
 
 # Update Room
+@login_required(login_url='login')
 def update_room(request, pk):
+    
     # Get the room we need to update by ID
     room = Room.objects.get(id=pk)
+    
     # Instantiate a form class (make a form object prefilled with room)
     form = RoomForm(instance=room)
+    
+    # Check if you are the owner of the room to have permission for updating
+    if request.user != room.host:
+        return HttpResponse('You are not allowed to update the room')
+    
     # Check if the method is POST
     if request.method == 'POST':
         form = RoomForm(request.POST, instance=room)
+        
         # Check if the form fields are valid, save them in the database
         if form.is_valid:
             form.save()
@@ -63,9 +120,15 @@ def update_room(request, pk):
     return render(request, 'base/room_form.html', context)
 
 # Delete Room
+@login_required(login_url='login')
 def delete_room(request, pk):
+    
     # Get the room by ID to delete
     room = Room.objects.get(id=pk)
+    
+    # Check if you are the owner of the room to have permission for deleting
+    if request.user != room.host:
+        return HttpResponse('You are not allowd to delete the room')
     # Check if the request is POST then delete the room
     if request.method == 'POST':
         room.delete()
